pyproject.toml
[project]
name = "latex-research-agent"
version = "0.1.0"
description = "Generate academic research papers with full LaTeX + PDF output"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "ray[serve]>=2.9.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.25.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "kodosumi",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
